# design/openapi.yaml
openapi: 3.0.0
info:
  title: Soroco House — Restaurant Order Management API
  version: 1.0.0
  description: |
    Backend API for the Soroco House QR-menu ordering system.

    Handles:
    - Customer flows: menu browsing, order placement with sandbox/mock payment,
      bill email.
    - Staff flows: login/signup (JWT with role), kitchen order board (FIFO),
      live updates over WebSocket.
    - Admin flows: menu CRUD, image upload, employee management, order history
      + CSV export.

    Conventions:
    - All JSON routes are under /api/v1 (this server URL). Exceptions: GET /health
      (raw health object) and the static /images mount.
    - Every JSON success uses the envelope {"success": true, "data": ...};
      every failure uses {"success": false, "message", "code", "detail"}.
    - Enums are stored as plain VARCHAR values (employee/admin, phonepay/razorpay,
      success/failed/cancelled, in_queue/preparing/prepared/delivered,
      ordered/delivered, standard/small/large).
    - Menu create/delete and image upload are ADMIN only; menu PATCH and kitchen
      routes allow employee + admin.
    - Prices are always recomputed server-side at POST /payments (5% tax).
    - Rate limits: auth 5/min, upload 5/min, payments 20/min.
    - Real-time events stream over WebSockets (documented in design/api-docs.md):
      ws /api/v1/ws/menu (public, frame {"type":"menu_updated"}) and
      ws /api/v1/ws/orders?token=<JWT> (frame {"type":"orders_updated","data":[...]}).
...
servers:
...
paths:
...
  /auth/login:
    post:
      summary: Staff login
      operationId: login
      tags: [Auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            examples:
              admin:
                value: { email: "thameem@test.com", password: "admin123" }
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
              examples:
                ok:
                  value:
                    success: true
                    data:
                      access_token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                      token_type: "bearer"
                      account_uuid: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
                      name: "Thameem"
                      email: "thameem@test.com"
                      role: "admin"
        '401':
          $ref: '#/components/responses/unauthorized'
        '429':
          $ref: '#/components/responses/rate_limited'
  /auth/signup:
    post:
      summary: Complete staff signup (set password)
      description: |
        Pre-created staff accounts have a NULL password. Signup accepts the
        registered email and sets the password (hashed server-side). On success
        a JWT (with the account role) is returned. Password confirm is
        client-side only and is NOT part of the payload.
      operationId: signup
      tags: [Auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignupRequest'
            examples:
              valid:
                value: { email: "employee@test.com", password: "strongpass1", confirm: "strongpass1" }
      responses:
        '200':
          description: Password set, JWT issued (same shape as login)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
              examples:
                ok:
                  value:
                    success: true
                    data:
                      access_token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                      token_type: "bearer"
                      account_uuid: "e5f6g7h8-9012-3456-7890-abcdef123456"
                      name: "Soroco Employee"
                      email: "employee@test.com"
                      role: "employee"
        '404':
          $ref: '#/components/responses/not_found'
        '409':
          $ref: '#/components/responses/conflict'
        '429':
          $ref: '#/components/responses/rate_limited'
  /auth/check:
    get:
      summary: Check whether an email is a staff account (signup step 1)
      operationId: checkAccount
      tags: [Auth]
      security: []
      parameters:
        - name: email
          in: query
          required: true
          schema:
            type: string
            format: email
            example: "employee@test.com"
      responses:
        '200':
          description: Always 200 — reports existence and password state.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
              examples:
                exists_without_password:
                  value:
                    success: true
                    data:
                      exists: true
                      has_password: false
        '429':
          $ref: '#/components/responses/rate_limited'
  /auth/me:
    get:
      summary: Current authenticated user
      operationId: getCurrentUser
      tags: [Auth]
      security:
        - bearer_auth: []
      responses:
        '200':
          description: Current user + role
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
              examples:
                ok:
                  value:
                    success: true
                    data:
                      account_uuid: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
                      name: "Thameem"
                      email: "thameem@test.com"
                      role: "admin"
        '401':
          $ref: '#/components/responses/unauthorized'

  # ─── Menu ───
...
  /admin/employees:
    get:
      summary: List all staff accounts
      operationId: listEmployees
      tags: [Admin]
      security:
        - bearer_auth: []
      responses:
        '200':
          description: Array of accounts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
              examples:
                ok:
                  value:
                    success: true
                    data:
                      - account_uuid: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
                        account_id: 1
                        account_name: "Thameem"
                        account_email: "thameem@test.com"
                        account_role: "admin"
                        created_at: "2026-09-20T09:00:00+00:00"
                        has_password: true
                        can_delete: false
                      - account_uuid: "e5f6g7h8-9012-3456-7890-abcdef123456"
                        account_id: 2
                        account_name: "Soroco Employee"
                        account_email: "employee@test.com"
                        account_role: "employee"
                        created_at: "2026-09-20T09:00:00+00:00"
                        has_password: false
                        can_delete: true
        '401':
          $ref: '#/components/responses/unauthorized'
        '403':
          $ref: '#/components/responses/forbidden'

    post:
      summary: Create a staff account
      description: Creates an employee account with NULL password (they sign up later).
      operationId: createEmployee
      tags: [Admin]
      security:
        - bearer_auth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmployeeCreate'
            examples:
              valid:
                value: { name: "New Staff", email: "new@test.com" }
      responses:
        '200':
          description: Account created (200 — not 201)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
              examples:
                created:
                  value:
                    success: true
                    data:
                      account_uuid: "f6f7f8f9-0123-4567-8901-abcdef123456"
                      account_id: 3
                      account_name: "New Staff"
                      account_email: "new@test.com"
                      account_role: "employee"
                      created_at: "2026-09-20T10:00:00+00:00"
                      has_password: false
                      can_delete: true
        '401':
          $ref: '#/components/responses/unauthorized'
        '403':
          $ref: '#/components/responses/forbidden'
        '409':
          $ref: '#/components/responses/conflict'
        '422':
          $ref: '#/components/responses/validation_error'
  /admin/employees/{account_uuid}:
    patch:
      summary: Update a staff account (name / email)
      operationId: updateEmployee
      tags: [Admin]
      security:
        - bearer_auth: []
      parameters:
        - name: account_uuid
          in: path
          required: true
          schema:
            type: string
            format: uuid
            example: "e5f6g7h8-9012-3456-7890-abcdef123456"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmployeePatch'
            examples:
              valid:
                value: { name: "Soroco Staff", email: "staff@test.com" }
      responses:
        '200':
          description: Updated account
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
        '401':
          $ref: '#/components/responses/unauthorized'
        '403':
          $ref: '#/components/responses/forbidden'
        '404':
          $ref: '#/components/responses/not_found'
        '409':
          $ref: '#/components/responses/conflict'
        '422':
          $ref: '#/components/responses/validation_error'

    delete:
      summary: Delete a staff account
      description: An admin account can never be deleted through this endpoint.
      operationId: deleteEmployee
      tags: [Admin]
      security:
        - bearer_auth: []
      parameters:
        - name: account_uuid
          in: path
          required: true
          schema:
            type: string
            format: uuid
            example: "e5f6g7h8-9012-3456-7890-abcdef123456"
      responses:
        '200':
          description: Account removed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessMessageEnvelope'
              examples:
                deleted:
                  value:
                    success: true
                    data:
                      message: "Employee removed"
        '401':
          $ref: '#/components/responses/unauthorized'
        '403':
          $ref: '#/components/responses/forbidden'
        '404':
          $ref: '#/components/responses/not_found'
        '409':
          $ref: '#/components/responses/conflict'

  # ─── Admin: Order history ───
  /admin/orders:
    get:
      summary: List orders by date range + status
      description: |
        Dates use YYYY-MM-DD. to_date is inclusive (end of day).
        order_status: ordered | delivered (omit for all).
      operationId: getOrderHistory
      tags: [Admin]
      security:
        - bearer_auth: []
      parameters:
        - name: from_date
          in: query
          required: false
          description: Range start (YYYY-MM-DD).
          schema:
            type: string
            format: date
            example: "2026-09-01"
        - name: to_date
          in: query
          required: false
          description: Range end, inclusive (YYYY-MM-DD).
          schema:
            type: string
            format: date
            example: "2026-09-20"
        - name: order_status
          in: query
          required: false
          schema:
            type: string
            enum: [ordered, delivered]
            example: "delivered"
      responses:
        '200':
          description: Array of serialized orders (same shape as kitchen list)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
              examples:
                ok:
                  value:
                    success: true
                    data: []
        '401':
          $ref: '#/components/responses/unauthorized'
        '403':
          $ref: '#/components/responses/forbidden'
        '422':
          $ref: '#/components/responses/validation_error'
  /admin/orders/export:
    get:
      summary: Export order history as CSV
      description: |
        Same filters as GET /admin/orders. Returns a text/csv attachment
        (NOT the JSON envelope). Columns: Order Number, Ordered At, Table,
        Customer, Items, Quantity, Subtotal (INR), Tax (INR), Total (INR),
        Payment Method, Payment Status, Kitchen Status, Order Status.
      operationId: exportOrderHistoryCsv
      tags: [Admin]
      security:
        - bearer_auth: []
      parameters:
        - name: from_date
          in: query
          required: false
          schema:
            type: string
            format: date
            example: "2026-09-01"
        - name: to_date
          in: query
          required: false
          schema:
            type: string
            format: date
            example: "2026-09-20"
        - name: order_status
          in: query
          required: false
          schema:
            type: string
            enum: [ordered, delivered]
            example: "delivered"
      responses:
        '200':
          description: CSV file download
          headers:
            Content-Disposition:
              description: 'attachment; filename="soroco-order-history.csv"'
              schema:
                type: string
          content:
            text/csv:
              schema:
                type: string
              examples:
                csv:
                  value: |
                    Order Number,Ordered At,Table,Customer,Items,Quantity,Subtotal (INR),Tax (INR),Total (INR),Payment Method,Payment Status,Kitchen Status,Order Status
                    12,2026-09-20T10:36:12.123456+00:00,A1,Alex,"Signature Cold Brew (large) x2; Darjeeling First Flush x1",3,460.00,23.00,483.00,razorpay,success,in_queue,ordered
        '401':
          $ref: '#/components/responses/unauthorized'
        '403':
          $ref: '#/components/responses/forbidden'
        '422':
          $ref: '#/components/responses/validation_error'

  # ─── WebSockets (documentation only — OpenAPI cannot express WS) ───
...
components:
  securitySchemes:
    bearer_auth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT issued by POST /auth/login or POST /auth/signup.
  responses:
    unauthorized:
      description: Missing, invalid or expired token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          examples:
            bad_token:
              value:
                success: false
                message: "Not authenticated"
                code: "UNAUTHORIZED"
                detail: null
    forbidden:
      description: Role insufficient for this route
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          examples:
            role:
              value:
                success: false
                message: "You don't have permission to perform this action"
                code: "FORBIDDEN"
                detail: null
    not_found:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          examples:
            missing:
              value:
                success: false
                message: "Menu item not found"
                code: "NOT_FOUND"
                detail: null
    conflict:
      description: Duplicate / forbidden-state conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          examples:
            dup:
              value:
                success: false
                message: "An account already exists for that email"
                code: "CONFLICT"
                detail: null
    validation_error:
      description: Pydantic or business-rule validation failure
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          examples:
            invalid:
              value:
                success: false
                message: "phone_number: Enter a valid phone number (10 digits)"
                code: "VALIDATION_ERROR"
                detail: null
    rate_limited:
      description: SlowAPI rate limit exceeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          examples:
            limited:
              value:
                success: false
                message: "Too many attempts. Please wait and try again."
                code: "RATE_LIMITED"
                detail: null
...
  schemas:
    SuccessEnvelope:
      type: object
      description: Standard success envelope — data carries the payload.
      required: [success, data]
      properties:
        success:
          type: boolean
          const: true
          example: true
        data:
          nullable: true
          description: Payload (object, array, string or number depending on endpoint).
    SuccessMessageEnvelope:
      type: object
      description: Success envelope for message-only responses (ok_message helper).
      required: [success, data]
      properties:
        success:
          type: boolean
          const: true
          example: true
        data:
          type: object
          required: [message]
          properties:
            message:
              type: string
              example: "Employee removed"
    ErrorEnvelope:
      type: object
      description: Standard error envelope for every non-2xx JSON response.
      required: [success, message, code, detail]
      properties:
        success:
          type: boolean
          const: false
          example: false
        message:
          type: string
          example: "Menu item not found"
        code:
          type: string
          nullable: true
          enum: [UNAUTHORIZED, INVALID_CREDENTIALS, FORBIDDEN, SIGNUP_PENDING, NOT_FOUND, CONFLICT, VALIDATION_ERROR, RATE_LIMITED, INTERNAL_ERROR, FORGED_CALLBACK]
          example: "NOT_FOUND"
        detail:
          type: string
          nullable: true
          example: null
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    SignupRequest:
      type: object
      required: [email, password, confirm]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        confirm:
          type: string
          description: Must equal password (validated server-side)
...
    EmployeeCreate:
      type: object
      required: [name, email]
      properties:
        name:
          type: string
        email:
          type: string
          format: email
    EmployeePatch:
      type: object
      description: At least one field required.
      properties:
        name:
          type: string
        email:
          type: string
          format: email
...